Correct Located URI visiblity
Located URIs should be visible when they exist at the search context org unit
or an ancestor thereof.  However, before the change embodied in this commit,
they effectively acted exactly like copies, making records visible anywhere
within the context+depth range.

An example, using the stock example OU hierarchy:

 * Before this change, a Located URI with an owner of BR1 would cause its
record to show up in a search at (or scoped to via a depth parameter) BR1,
SYS1 or CONS.

 * Before this change, a Located URI with an owner of SYS1 would NOT cause its
record to show up in a search at (or scoped to via a depth parameter) BR1!

 * After this change, a Located URI with an owner of BR1 will cause its record
to show up only in a search with a context OU of BR1, irrespective of depth
scoping.

 * After this change, a Located URI with an owner of SYS1 WILL cause its
record to show up in a search with a context OU of BR1 or SYS1, regardless of
depth scoping.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Dan Scott <dscott@laurentian.ca>