Really add *entire* selection list to a purchase order, when desired,
and generally do what a user would expect when, in Acq, the user
executes an operation on "all items".

This addresses LaunchPad bug #810030.

The fundamental issue is that an AcqLiTable, which is used in several
different ways in different interfaces, may be representing items on a
selection list (for example), but it doesn't necessarily know all of the
items that belong to that selection list.  It only knows what it has drawn.

This addresses that deficiency by teaching the getSelected() method
some new tricks so that it can consult the server in the various ways
appropriate to getting a list of /all/ the items on a selection list, a
purchase order, a virtual PO, a set of search results, etc.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>