Z39.50 local catalog fetch-more repairs
Fixes the following problem:

When paging through local catalog results in the Z39.50 interface,
paging off then end of the local results would result in an unknown
server error caused by trying to retrieve an empty list of records.

$list->{count} in this context is the overall hit count, not the counts
in the current page of results.  Note also if {count} is zero, the list
of IDs will be undef, so we use both.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Jason Etheridge <jason@esilibrary.com>