Avoid a possible bug in selection list deletion
This particularly matters for the "ephemeral," unnamed selection list that
Acquisitions General Search uses to hold bib records represented as
lineitems, as that list is deleted for the next run of that kind of
search.

It's possible to get lineitems onto that list that are in a state
other than "new".  Code that deletes this list first deletes the
lineitems that belong to it, but that code formerly assumed that there
are two types of lineitems. "new" ones (that should be deleted) and ones
that belong to purchase orders (that should be deatched).  That doesn't
necessarily cover all possible lineitems on the picklist, so deletion
ultimately fails.

This fixes that.

[update: now this bugfix incorporates a bugfix, the need for which was
detected by Ben Shum. See LP #869581]

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Bill Erickson <berick@esilibrary.com>