TPac: Use workstation to determine staff
Instead of user agent, which may be set easily by anyone, use the presence
of a workstation ID to determine "staff" use of the TPac.

This is much less likely to be spoofed or incorrect due to things like
previous use of the extension version of the client in Firefox.

Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>
Signed-off-by: Bill Erickson <berick@esilibrary.com>