Give OrgTree.js a locale, always
We started generating locale-specific OrgTree.js files many moons ago,
and now no longer provide a "default" locale in the root of the
common/js directory. The result was that the index.xhtml page for the
staff client was throwing an "_l is undefined" exception; largely
meaningless, as it seems that the OrgTree isn't even used on that page,
but noise that doesn't need to be in the JavaScript console.

Here we correct that exception by providing a locale, and hopefully
prevent any further copy/paste mistakes.

Signed-off-by: Dan Scott <dscott@laurentian.ca>