TPac: bookbags - stop trying to show everything at once
Go back to displaying one bookbag at a time (not counting the temporary
list, when that is populated)

This hopefully constitutes a usability improvement.

While it doesn't provide paging of bookbags themselves or of the items
therein, it should make that easier to do.  Showing all bookbags and all
their items made it nonobvious how paging would even be expected to
behave.

This also provides the same sorting for the items in the temporary list
as in real bookbags.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Dan Scott <dscott@laurentian.ca>