LP875819: Fix SRU and Z39.50 default search index failures
Per http://www.loc.gov/standards/sru/specs/cql.html :

All parts of CQL are case insensitive [...] If any case insensitive
part of CQL is specified with both upper and lower case, it is
for aesthetic purposes only.

So fold the SRU qualifier and relation to lower case. This then
matches the srw.serverchoice entry that we already have in
config.metabib_search_alias so that default searches will successfully
point to the keyword index.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>