TPac: Put staff saved searches in same div as facets on the results page
And on the record page, make them less ugly. We need to get rid of this
974-px-wide business. It's ridiculous, and it keeps us busy fighting CSS
just to keep a silly grey margin on the right side of the display, and
it keeps us from using all available space.  We need to simplify the CSS
greatly, in fact. Let's bite that bullet sooner than later.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>