TPAC: Bring used CSS definitions over from semiauto.css
Convert ".red" class to ".alert", delete classes that were defined in
semiauto.css as being verified as being in use that are actually no longer
in use, and merge the remainder to style.css. This will avoid one more
HTTP request as well as a huge mess of CSS to have to worry about
carrying around.

Signed-off-by: Dan Scott <dscott@laurentian.ca>