ACQ: remove piles of unused cancel_reasons
There is a large set of of acq.cancel_reason's in Evergreen that come
from the EDI spec and will never be used in practice.  Since there are
scenarios where cancel reasons are manually chosen by staff, let's
remove all of the extras so the process of selecting a cancel reason is
not so fraught with "Poultry" and "Airline attendants", etc.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>