Make idlval build, rather than be dead code
The tool "idlval" contained in Open-ILS/src/c-apps can be used to check
the correctness of fieldmapper definitions in files like
Open-ILS/examples/fm_IDL.xml - but currently does not have a build
target, nor does it compile cleanly due to the use of unspecified
OpenSRF header files. This commit adds a build target, fixes the compile
problems, and installs the tool in the bin directory.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Mike Rylander <mrylander@gmail.com>