Tpac: fix the doubly wrapped holds editor
For some reason it must have made since in the past to wrap this
interface in both parts/base.tt2 and myopac/parts/base.tt2, but
nowadays that results in having extra headers and footers around
the actual page contents, so the layout looks broken.

This fixes that. No other interfaces were doing this double wrapping.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Bill Erickson <berick@esilibrary.com>