Tpac: Rename orig_loc to physical_loc, for clarity's sake.
This is an OPAC parameter not so much meant for use by patrons at home,
but for users at machines physically located in libraries.  In
conjunction with OpenILS::WWW::Redirector, it helps those machines
"remember" what branch they're in.

Anyway, the name orig_loc doesn't make the parameter's meaning quite as
clear as it could whne you encounter it in code.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Bill Erickson <berick@esilibrary.com>