Tpac; hold error message 18n string extraction repairs
Make it possible for xgettext.pl (using
Locale::Maketext::Extract::Plugin::TT2) to extract hold error
message strings for translation.  The plugin is not quite
smart enough to extract wholly dynamic values from the templates, since
it (presumably) only parses the templates and does not execute them.
The solution is to wrap the strings in l() where they are defined.

Add an example for the EVENT_MSG_MAP for the event PATRON_EXCEEDS_FINES.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>