TPac; hold success confirmation; redirect repairs
1. After a successful hold placement, provide feedback to the user that
the hold placement succeeded.  After placement, the user is taken to a
page (same page for failed holds) where the title(s) are listed along
with the success/failure information for each.

2. In some cases, the user may be redirected to the my account page
instead of the point of origin for holds placement.  This occurs
generally when the user is not already logged in and is asked to log in
prior to holds placement.

3. As a side effect, this change replaces one of the ornery
history.go(-1) actions with a true href/URL.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>