TPac; detail page "Search Results" page correction
On the detail page, when the user selects the "Search Results" link to
return the original search results, return to the correct page in the
search.  The correct page will be the page where the currently viewed
record sits within the search.  Normally, this is the last viewed
results page.  However, if the user does some paging within the detail
page, enough to cross a page boundry, "Search Results" will take them to
the new page (i.e. the page w/ the current record).

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>