Default to org tree root for bre unapi funcs
If the org unit ("site") parameter to unapi.biblio_record_entry_feed()
or unapi.bre() is equal to "-" or NULL, default to the root of the org
unit tree.  In cases where the caller does not care to limit by site,
determing the shortname of the root org node is one less step the client
now has to take.

Original patch from Mike Rylander.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>