LP887822 mint_condition/Quality not set correctly
during item creation.

Specifically, the Javascript object for a new item (acp) does not explicitly set
the mint_condition field, leaving it null. The item attribute editor treats the
null as false (which it displays as "Mediocre" for that field). The database has
a not null constraint the corresponding column in asset.copy, and defaults to
true. So a new item in the attribute editor will show "Mediocre", but when
actually created, will become "Good".

This branch just sets mint_condition to true when an item is created.

Signed-off-by: Jason Etheridge <jason@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>