Tpac; repair login link after password reset
After a successful password reset, the UI offers a link to "Log In to My
Account".  Loging in this way would ultimately redirect the user back to
the password reset form.  This patch updates the Log In link to point to
myopac/main, which will direct the user to log in first,  instead of
pointing directly to /login, which, lacking any other information,
will return the user to the referer after logging in.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>