TPac; disallow certain actions for captured holds
In the hold edit page, where users can update the pickup library,
suspend holds, etc. disallow actions that don't make sense for holds
that have already been captured.  These include suspending the hold and
changing the expire time.

This patch also replaces the confusing "Cancel" button in the hold edit
interface (which does not cancel the hold) with a "Go Back" button which
returns to the holds list.

Signed-off-by: Bill Erickson <berick@esilibrary.com>