Create a search_normalize() variant of naco_normalize()
In the quest to support searching text that contains leading articles
joined by an apostrophe - for example, "l'histoire" - such that a
searcher can enter either "l'histoire" or "histoire" and get results -
add a variant of naco_normalize() that does not strip the apostrophe
entirely, but rather replaces it with a space such that the root word
can be indexed appropriately.

This implementation refactors the OpenILS::Utils::Normalize code to make
the differences between search_normalize() and naco_normalize() as clear
as possible, but duplicates code significantly in the in-db version of
the code. Someday maybe the database can rely on
OpenILS::Utils::Normalize instead of inline functions :)

Signed-off-by: Dan Scott <dscott@laurentian.ca>