Index leading articles with apostrophes correctly
Titles with non-filing characters used to get separated by spaces; not a
problem in English with leading articles like "The" and "A" which are
followed by spaces anyway, but a significant problem for French where
"L'année" is not supposed to be turned into "L' année".

To combat this, we define a titleNonfiling variant in the MODS32
stylesheets that ignores the second indicator and does not wrap leading
articles in the nonSort element, avoiding the injection of unwanted
spaces.

Signed-off-by: Dan Scott <dscott@laurentian.ca>