LP825039: Add reingest logic for non-filing titles
Rather than reindexing all titles that have a non-filing indicator, this
upgrade handles the known problematic cases (apostrophes with leading
articles). On our system that's 55K out of 2.2M records - not bad at
all.

We also ensure the format is set to MODS32 for title:proper, as sites
might have customized it on a MARC basis and we don't want to reindex
the records with NULL values due to a mix of format and XPath.

Signed-off-by: Dan Scott <dscott@laurentian.ca>