Serials: in the holding code mini-wizard, offer a dropdown for seasons
This spares the user from having to remember the MFHD code for the
season, when dealing with periodicals issued by season (21 - 24 for
Spring - Winter, respectively).

This also applies sets the "scrollOnFocus" attribute to false for the
dijits within the mini wizard, as the jumpy behavior you get in
less-than-fullscreen staff client windows is otherwise maddening.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>