Reindex is not just for non-filing indicators anymore
Since we discovered the apostrophe parsing problem affects records with
text containing apostrophes at any position, don't limit ourselves to
non-filing indicator titles. In theory you might want to reindex any
record that contains an apostrophe, but in practice we found the most
confusion occurs with searching for titles - so target those in the
upgrade script.

Signed-off-by: Dan Scott <dscott@laurentian.ca>