TPac; support context-genereted refresh content
If the tpac context object contains a "refresh" attribute, it will be
used as the content attribute in a <meta http-equiv="refresh" ...>
element in the page head.  If present, it will override the default
refresh content for the authentication timeout.

While we're in there, let's use the more correct form of the content
attribute:  content="x; url=foo"

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>