Enable record summary to be scrollable at larger font sizes
The record summary at the top of the cataloguing interfaces appears to
determine its size based on the "normal" font size at load; then, when a
larger font size is applied to the page, the text gets hidden by the
fixed size of the groupbox.  Ideally the groupbox would adjust to
accomodate the larger font size, but after trying many different
approaches the best short-term approach seems to be to enable overflow
so that scrollbars can appear when necessary.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Jason Etheridge <jason@esilibrary.com>