LP893315: Do not return delected call numbers in holdings_xml
Per LP893315, deleted call numbers were obscuring the actual active URIs
associated with a given bib record in a given scope at the default limit
of 5 results per record in unapi.holdings_xml.

Filtering out the deleted call numbers from the results seems like a
good idea, as we're generally interested in only showing the active
results (at least as in-db unapi is currently used). This also addresses
the problem of not showing active URIs when there are active URIs to
display.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Bill Erickson <berick@esilibrary.com>