LP881774: Missing schema on function in 2.0-2.1 upgrade
Although the search path is altered to include the evergreen schema at
the start of the upgrade, one of the functions in the evergreen schema
does not resolve when invoked without the explicit schema later in the
upgrade script. We address that superficial (but failure provoking / time
wasting) problem by qualifying the function with the evergreen schema in
this commit, but we should also find out why the altered search path
does not seem to be functioning in the upgrade script.

Signed-off-by: Dan Scott <dscott@laurentian.ca>