TPAC: Use padding instead of explicit width for tabs
Explicit widths work well for a combination of a specific font sizes and
a known string; however, in the context of translation, strings may get
longer than the explicit width allows. Using padding gives us more
flexibility to accommodate different font sizes and variable length
strings; the uniformity of the tab size will be sacrificed but perhaps
that is not such a bad thing.

Signed-off-by: Dan Scott <dscott@laurentian.ca>