TPAC: Remove more explicit widths from results page
Refactor the facet sidebar such that headers consist of a floated span
and a div with no specific width; absent a specific width, the button
will float over top of text that might expand due to translation.

Also remove the explicit table cell width for the results counter column
and move some inline styles over to CSS. More to be done in results
table but we are on our way.

Signed-off-by: Dan Scott <dscott@laurentian.ca>