Add concerto.sql dataset with callnums, copies, and URIs
Useful for testing a fresh Evergreen database, the concerto.sql dataset
is an SQL file that adds 100 bib records and automatically generates
assorted call numbers and copies for 3 of the branches. 50 of the
records include unadorned 856 fields and 25 of them have located URIs
(25 scoped at SYS1, and 10 scoped at both SYS1 and BR3).

In the effort to make the one-shot load of sample bib / copy / call
number data broad and realistic, delete some of the entities, add
prefixes and suffixes to some of the call numbers, and include some
monograph parts and conjoined items.

Signed-off-by: Dan Scott <dan@coffeecode.net>
Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>