TPAC: Fix record summary e-resource and copy display
Electronic resources were wrapped with an ugly black line and no
differentiation from the outstanding text. Give them a light background
and remove the black line. Also, use an unordered list if we have more
than one URI to display.

Move the copy summary below the "Copies" heading that immediately
follows it for a more logical grouping of elements.

Signed-off-by: Dan Scott <dscott@laurentian.ca>