TPAC: results use floated DIVs instead of table layout
Rather than a table containing a single row for the rather gross purpose
of placing the facet sidebar on the left, a little bit of a divider
between the facets and the results, and the results on the right, just
use a floated div and a relatively positioned div for the two pertinent
chunks of content. Bonus: one touch of CSS moves the facet sidebar to
the right instead of the left.

Signed-off-by: Dan Scott <dscott@laurentian.ca>