Process still-shelved, canceled holds in clear-shelf process
Holds that are canceled but still physically on the hold shelf are
processed like other holds during the clear-shelf process.  They will be
grouped in the same manner and if a custom post-clear status is
configured, copies linked to already-canceled holds will get the custom
status.  The only difference is already-canceled holds will not be
re-canceled during the clear-shelf process.

This change has the added benefit of leveraging another API call to
retrieve the shelf-expired holds instead of using its own custom
json-query.

See https://bugs.launchpad.net/evergreen/+bug/904332

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>