Add a .gitignore file suitable for master
"git status" is painful in a source directory in which you're built and
installed Evergreen. This commit adds a .gitignore file that masks most
of the files and directories that you don't want to even think about, in
the hopes that the default Evergreen development experience can be mo'
betta.

Variations on this could be backported to rel_2_1 and rel_2_0 but take
care as some of the files that used to be source files in older releases
are now build from *.in files via autoconf.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>